Move objective_function to preview section
